+++ src/agenda/index.js
@@ -1,16 +1,20 @@
 import _ from 'lodash';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
+import memoize from 'memoize-one';
+
 import React, {Component} from 'react';
 import {Text, View, Dimensions, Animated} from 'react-native';
+
 import {extractComponentProps} from '../component-updater';
-import {parseDate, xdateToData} from '../interface';
+import {parseDate, xdateToData, toMarkingFormat} from '../interface';
 import dateutils from '../dateutils';
 import {AGENDA_CALENDAR_KNOB} from '../testIDs';
 import {VelocityTracker} from '../input';
 import styleConstructor from './style';
 import CalendarList from '../calendar-list';
 import ReservationList from './reservation-list';
+
 
 const HEADER_HEIGHT = 104;
 const KNOB_HEIGHT = 24;
@@ -178,22 +182,19 @@ export default class AgendaView extends Component {
     _.invoke(this.props, 'onDayPress', xdateToData(day));
   }
 
-  generateMarkings = () => {
-    const {markedDates, items} = this.props;
-    let markings = markedDates;
-
-    if (!markings) {
-      markings = {};
-      Object.keys(items || {}).forEach(key => {
+  generateMarkings = memoize((selectedDay, markedDates, items = {}) => {
+    if (!markedDates) {
+      markedDates = {};
+      Object.keys(items).forEach(key => {
         if (items[key] && items[key].length) {
-          markings[key] = {marked: true};
+          markedDates[key] = {marked: true};
         }
       });
     }
 
-    const key = this.state.selectedDay.toString('yyyy-MM-dd');
-    return {...markings, [key]: {...(markings[key] || {}), ...{selected: true}}};
-  };
+    const key = toMarkingFormat(selectedDay);
+    return {...markedDates, [key]: {...(markedDates[key] || {}), ...{selected: true}}};
+  });
 
   onScrollPadLayout = () => {
     // When user touches knob, the actual component that receives touch events is a ScrollView.
@@ -297,6 +298,7 @@ export default class AgendaView extends Component {
   }
 
   renderCalendarList() {
+    const {markedDates, items} = this.props;
     const shouldHideExtraDays = this.state.calendarScrollable ? this.props.hideExtraDays : false;
     const calendarListProps = extractComponentProps(CalendarList, this.props);
 
@@ -305,7 +307,7 @@ export default class AgendaView extends Component {
         {...calendarListProps}
         ref={c => (this.calendar = c)}
         current={this.currentMonth}
-        markedDates={this.generateMarkings()}
+        markedDates={this.generateMarkings(this.state.selectedDay, markedDates, items)}
         calendarWidth={this.viewWidth}
         scrollEnabled={this.state.calendarScrollable}
         hideExtraDays={shouldHideExtraDays}
@@ -331,10 +333,22 @@ export default class AgendaView extends Component {
     return knob;
   }
 
+  renderWeekDaysNames = memoize((weekDaysNames) => {    
+    return weekDaysNames.map((day, index) => (
+      <Text allowFontScaling={false} key={day + index} style={this.style.weekday} numberOfLines={1}>
+        {day}
+      </Text>
+    ));
+  });
+
+  renderWeekNumbersSpace = () => {
+    return this.props.showWeekNumbers && <View allowFontScaling={false} style={this.style.weekday} numberOfLines={1}/>;
+  };
+
   render() {
-    const {firstDay, hideKnob, showWeekNumbers, style, testID} = this.props;
-    const agendaHeight = this.initialScrollPadPosition();
+    const {firstDay, hideKnob, style, testID} = this.props;
     const weekDaysNames = dateutils.weekDayNames(firstDay);
+    const agendaHeight = this.initialScrollPadPosition();
     const weekdaysStyle = [
       this.style.weekdays,
       {
@@ -390,21 +404,17 @@ export default class AgendaView extends Component {
     };
 
     return (
-      <View testID={testID} onLayout={this.onLayout} style={[style, {flex: 1, overflow: 'hidden'}]}>
+      <View testID={testID} onLayout={this.onLayout} style={[style, this.style.container]}>
         <View style={this.style.reservations}>{this.renderReservations()}</View>
         <Animated.View style={headerStyle}>
-          <Animated.View style={{flex: 1, transform: [{translateY: contentTranslate}]}}>
+          <Animated.View style={[this.style.animatedContainer, {transform: [{translateY: contentTranslate}]}]}>
             {this.renderCalendarList()}
           </Animated.View>
           {this.renderKnob()}
         </Animated.View>
         <Animated.View style={weekdaysStyle}>
-          {showWeekNumbers && <Text allowFontScaling={false} style={this.style.weekday} numberOfLines={1}></Text>}
-          {weekDaysNames.map((day, index) => (
-            <Text allowFontScaling={false} key={day + index} style={this.style.weekday} numberOfLines={1}>
-              {day}
-            </Text>
-          ))}
+          {this.renderWeekNumbersSpace()}
+          {this.renderWeekDaysNames(weekDaysNames)}
         </Animated.View>
         <Animated.ScrollView
           ref={ref => (this.scrollPad = ref)}