+++ src/calendar/header/index.js
@@ -1,8 +1,11 @@
 import _ from 'lodash';
 import PropTypes from 'prop-types';
+import memoize from 'memoize-one';
 import XDate from 'xdate';
+
 import React, {Component, Fragment} from 'react';
 import {ActivityIndicator, Platform, View, Text, TouchableOpacity, Image} from 'react-native';
+
 import {shouldUpdate} from '../../component-updater';
 import {weekDayNames} from '../../dateutils';
 import {
@@ -13,6 +16,7 @@ import {
   HEADER_MONTH_NAME
 } from '../../testIDs';
 import styleConstructor from './style';
+
 
 class CalendarHeader extends Component {
   static displayName = 'IGNORE';
@@ -105,7 +109,7 @@ class CalendarHeader extends Component {
     return this.addMonth();
   };
 
-  renderWeekDays = weekDaysNames => {
+  renderWeekDays = memoize(weekDaysNames => {
     const {disabledDaysIndexes} = this.props;
 
     return weekDaysNames.map((day, idx) => {
@@ -115,13 +119,17 @@ class CalendarHeader extends Component {
         dayStyle.push(this.style.disabledDayHeader);
       }
 
+      if (this.style[`dayTextAtIndex${idx}`]) {
+        dayStyle.push(this.style[`dayTextAtIndex${idx}`]);
+      }
+
       return (
         <Text allowFontScaling={false} key={idx} style={dayStyle} numberOfLines={1} accessibilityLabel={''}>
           {day}
         </Text>
       );
     });
-  };
+  });
 
   renderHeader = () => {
     const {renderHeader, month, monthFormat, testID, webAriaLevel} = this.props;