+++ src/calendar/index.js
@@ -1,12 +1,16 @@
 import _ from 'lodash';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
+import memoize from 'memoize-one';
+
 import React, {Component} from 'react';
 import {View} from 'react-native';
 import GestureRecognizer, {swipeDirections} from 'react-native-swipe-gestures';
+
 import dateutils from '../dateutils';
-import {xdateToData, parseDate} from '../interface';
-import shouldComponentUpdate from './updater';
+import {xdateToData, parseDate, toMarkingFormat} from '../interface';
+import {getState} from '../day-state-manager';
+// import shouldComponentUpdate from './updater';
 import {extractComponentProps} from '../component-updater';
 import {WEEK_NUMBER} from '../testIDs';
 import styleConstructor from './style';
@@ -14,7 +18,6 @@ import CalendarHeader from './header';
 import BasicDay from './day/basic';
 import Day from './day/index';
 
-const EmptyArray = [];
 
 /**
  * @description: Calendar component
@@ -82,7 +85,7 @@ class Calendar extends Component {
       currentMonth: props.current ? parseDate(props.current) : XDate()
     };
 
-    this.shouldComponentUpdate = shouldComponentUpdate;
+    // this.shouldComponentUpdate = shouldComponentUpdate;
   }
   
   addMonth = count => {
@@ -94,11 +97,7 @@ class Calendar extends Component {
       return;
     }
 
-    this.setState(
-      {
-        currentMonth: day.clone()
-      },
-      () => {
+    this.setState({currentMonth: day.clone()}, () => {
         if (!doNotTriggerListeners) {
           const currMont = this.state.currentMonth.clone();
           _.invoke(this.props, 'onMonthChange', xdateToData(currMont));
@@ -108,7 +107,7 @@ class Calendar extends Component {
     );
   };
 
-  _handleDayInteraction(date, interaction) {
+  handleDayInteraction(date, interaction) {
     const {disableMonthChange} = this.props;
     const day = parseDate(date);
     const minDate = parseDate(this.props.minDate);
@@ -127,46 +126,14 @@ class Calendar extends Component {
   }
 
   pressDay = date => {
-    this._handleDayInteraction(date, this.props.onDayPress);
+    this.handleDayInteraction(date, this.props.onDayPress);
   };
 
   longPressDay = date => {
-    this._handleDayInteraction(date, this.props.onDayLongPress);
+    this.handleDayInteraction(date, this.props.onDayLongPress);
   };
 
-  getDateMarking(day) {
-    const {markedDates} = this.props;
-
-    if (!markedDates) {
-      return false;
-    }
-
-    const dates = markedDates[day.toString('yyyy-MM-dd')] || EmptyArray;
-
-    if (dates.length || dates) {
-      return dates;
-    } else {
-      return false;
-    }
-  }
-
-  getState(day) {
-    const {disabledByDefault} = this.props;
-    const minDate = parseDate(this.props.minDate);
-    const maxDate = parseDate(this.props.maxDate);
-    let state = '';
-
-    if (disabledByDefault) {
-      state = 'disabled';
-    } else if (dateutils.isDateNotInTheRange(minDate, maxDate, day)) {
-      state = 'disabled';
-    } else if (!dateutils.sameMonth(day, this.state.currentMonth)) {
-      state = 'disabled';
-    } else if (dateutils.sameDate(day, XDate())) {
-      state = 'today';
-    }
-    return state;
-  }
+  swipeProps = {onSwipe: (direction, state) => this.onSwipe(direction, state)};
 
   onSwipe = gestureName => {
     const {SWIPE_UP, SWIPE_DOWN, SWIPE_LEFT, SWIPE_RIGHT} = swipeDirections;
@@ -192,13 +159,13 @@ class Calendar extends Component {
     this.header.onPressLeft();
   };
 
-  renderWeekNumber(weekNumber) {
+  renderWeekNumber = memoize((weekNumber) => {
     return (
       <View style={this.style.dayContainer} key={`week-container-${weekNumber}`}>
         <BasicDay
           key={`week-${weekNumber}`}
-          marking={{disableTouchEvent: true}}
-          state="disabled"
+          marking={{disabled: true, disableTouchEvent: true}}
+          // state='disabled'
           theme={this.props.theme}
           testID={`${WEEK_NUMBER}-${weekNumber}`}
         >
@@ -206,10 +173,10 @@ class Calendar extends Component {
         </BasicDay>
       </View>
     );
-  }
+  });
 
   renderDay(day, id) {
-    const {hideExtraDays} = this.props;
+    const {hideExtraDays, markedDates} = this.props;
     const dayProps = extractComponentProps(Day, this.props);
 
     if (!dateutils.sameMonth(day, this.state.currentMonth) && hideExtraDays) {
@@ -221,8 +188,8 @@ class Calendar extends Component {
         <Day
           {...dayProps}
           day={day}
-          state={this.getState(day)}
-          marking={this.getDateMarking(day)}
+          state={getState(day, this.state.currentMonth, this.props)}
+          marking={markedDates?.[toMarkingFormat(day)]}
           onPress={this.pressDay}
           onLongPress={this.longPressDay}
         />
@@ -268,34 +235,33 @@ class Calendar extends Component {
     let indicator;
 
     if (current) {
-      const lastMonthOfDay = current.clone().addMonths(1, true).setDate(1).addDays(-1).toString('yyyy-MM-dd');
+      const lastMonthOfDay = toMarkingFormat(current.clone().addMonths(1, true).setDate(1).addDays(-1));
       if (displayLoadingIndicator && !(markedDates && markedDates[lastMonthOfDay])) {
         indicator = true;
       }
     }
 
     const headerProps = extractComponentProps(CalendarHeader, this.props);
-
-    const props = {
-      ...headerProps,
-      testID: testID,
-      style: headerStyle,
-      ref: c => (this.header = c),
-      month: this.state.currentMonth,
-      addMonth: this.addMonth,
-      displayLoadingIndicator: indicator
-    };
-
     const CustomHeader = customHeader;
     const HeaderComponent = customHeader ? CustomHeader : CalendarHeader;
 
-    return <HeaderComponent {...props} />;
+    return (
+      <HeaderComponent 
+        {...headerProps}
+        testID={testID}
+        style={headerStyle}
+        ref={r => this.header = r}
+        month={this.state.currentMonth}
+        addMonth={this.addMonth}
+        displayLoadingIndicator={indicator}
+      />
+    );
   }
 
   render() {
     const {enableSwipeMonths, style} = this.props;
     const GestureComponent = enableSwipeMonths ? GestureRecognizer : View;
-    const gestureProps = enableSwipeMonths ? {onSwipe: (direction, state) => this.onSwipe(direction, state)} : {};
+    const gestureProps = enableSwipeMonths ? this.swipeProps : undefined;
 
     return (
       <GestureComponent {...gestureProps}>