+++ src/expandableCalendar/calendarProvider.js
@@ -1,11 +1,12 @@
 import _ from 'lodash';
-import React, {Component} from 'react';
-import {Animated, TouchableOpacity, View} from 'react-native';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
 
+import React, {Component} from 'react';
+import {StyleSheet, Animated, TouchableOpacity, View} from 'react-native';
+
 import dateutils from '../dateutils';
-import {xdateToData} from '../interface';
+import {xdateToData, toMarkingFormat} from '../interface';
 import styleConstructor from './style';
 import CalendarContext from './calendarContext';
 
@@ -38,14 +39,15 @@ class CalendarProvider extends Component {
     todayButtonStyle: PropTypes.oneOfType([PropTypes.object, PropTypes.number, PropTypes.array]),
     /** The opacity for the disabled today button (0-1) */
     disabledOpacity: PropTypes.number
-  }
+  };
 
   constructor(props) {
     super(props);
     this.style = styleConstructor(props.theme);
 
     this.state = {
-      date: this.props.date || XDate().toString('yyyy-MM-dd'),
+      prevDate: this.props.date || toMarkingFormat(XDate()),
+      date: this.props.date || toMarkingFormat(XDate()),
       updateSource: UPDATE_SOURCES.CALENDAR_INIT,
       buttonY: new Animated.Value(-props.todayBottomMargin || -TOP_POSITION),
       buttonIcon: this.getButtonIcon(props.date),
@@ -64,6 +66,7 @@ class CalendarProvider extends Component {
     return {
       setDate: this.setDate,
       date: this.state.date,
+      prevDate: this.state.prevDate,
       updateSource: this.state.updateSource,
       setDisabled: this.setDisabled
     };
@@ -72,7 +75,7 @@ class CalendarProvider extends Component {
   setDate = (date, updateSource) => {
     const sameMonth = dateutils.sameMonth(XDate(date), XDate(this.state.date));
 
-    this.setState({date, updateSource, buttonIcon: this.getButtonIcon(date)}, () => {
+    this.setState({date, prevDate: this.state.date, updateSource, buttonIcon: this.getButtonIcon(date)}, () => {
       this.animateTodayButton(date);
     });
 
@@ -81,14 +84,14 @@ class CalendarProvider extends Component {
     if (!sameMonth) {
       _.invoke(this.props, 'onMonthChange', xdateToData(XDate(date)), updateSource);
     }
-  }
+  };
 
-  setDisabled = (disabled) => {
+  setDisabled = disabled => {
     if (this.props.showTodayButton && disabled !== this.state.disabled) {
       this.setState({disabled});
       this.animateOpacity(disabled);
     }
-  }
+  };
 
   getButtonIcon(date) {
     if (!this.props.showTodayButton) {
@@ -120,7 +123,7 @@ class CalendarProvider extends Component {
 
   animateTodayButton(date) {
     if (this.props.showTodayButton) {
-      const today = XDate().toString('yyyy-MM-dd');
+      const today = toMarkingFormat(XDate());
       const isToday = today === date;
 
       Animated.spring(this.state.buttonY, {
@@ -144,9 +147,9 @@ class CalendarProvider extends Component {
   }
 
   onTodayPress = () => {
-    const today = XDate().toString('yyyy-MM-dd');
+    const today = toMarkingFormat(XDate());
     this.setDate(today, UPDATE_SOURCES.TODAY_PRESS);
-  }
+  };
 
   renderTodayButton() {
     const {disabled, opacity, buttonY, buttonIcon} = this.state;
@@ -155,9 +158,15 @@ class CalendarProvider extends Component {
 
     return (
       <Animated.View style={[this.style.todayButtonContainer, {transform: [{translateY: buttonY}]}]}>
-        <TouchableOpacity style={[this.style.todayButton, this.props.todayButtonStyle]} onPress={this.onTodayPress} disabled={disabled}>
-          <Animated.Image style={[this.style.todayButtonImage, {opacity}]} source={buttonIcon}/>
-          <Animated.Text allowFontScaling={false} style={[this.style.todayButtonText, {opacity}]}>{today}</Animated.Text>
+        <TouchableOpacity
+          style={[this.style.todayButton, this.props.todayButtonStyle]}
+          onPress={this.onTodayPress}
+          disabled={disabled}
+        >
+          <Animated.Image style={[this.style.todayButtonImage, {opacity}]} source={buttonIcon} />
+          <Animated.Text allowFontScaling={false} style={[this.style.todayButtonText, {opacity}]}>
+            {today}
+          </Animated.Text>
         </TouchableOpacity>
       </Animated.View>
     );
@@ -166,9 +175,7 @@ class CalendarProvider extends Component {
   render() {
     return (
       <CalendarContext.Provider value={this.getProviderContextValue()}>
-        <View style={[{flex: 1}, this.props.style]}>
-          {this.props.children}
-        </View>
+        <View style={[styles.container, this.props.style]}>{this.props.children}</View>
         {this.props.showTodayButton && this.renderTodayButton()}
       </CalendarContext.Provider>
     );
@@ -176,3 +183,9 @@ class CalendarProvider extends Component {
 }
 
 export default CalendarProvider;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1
+  }
+});