+++ src/expandableCalendar/index.js
@@ -1,18 +1,21 @@
 import _ from 'lodash';
+import PropTypes from 'prop-types';
+import memoize from 'memoize-one';
+import XDate from 'xdate';
+
 import React, {Component} from 'react';
 import {AccessibilityInfo, PanResponder, Animated, View, Text, Image} from 'react-native';
-import PropTypes from 'prop-types';
-import XDate from 'xdate';
-import {CALENDAR_KNOB} from '../testIDs';
 
+import {CALENDAR_KNOB} from '../testIDs';
 import dateutils from '../dateutils';
-import {parseDate} from '../interface';
-import styleConstructor from './style';
+import {parseDate, toMarkingFormat} from '../interface';
+import styleConstructor, {HEADER_HEIGHT} from './style';
 import CalendarList from '../calendar-list';
 import Calendar from '../calendar';
 import asCalendarConsumer from './asCalendarConsumer';
 import WeekCalendar from './weekCalendar';
 import Week from './week';
+
 
 const commons = require('./commons');
 const UPDATE_SOURCES = commons.UPDATE_SOURCES;
@@ -25,8 +28,8 @@ const BOUNCINESS = 6;
 const CLOSED_HEIGHT = 120; // header + 1 week
 const WEEK_HEIGHT = 46;
 const KNOB_CONTAINER_HEIGHT = 20;
-const HEADER_HEIGHT = 68;
 const DAY_NAMES_PADDING = 24;
+const PAN_GESTURE_THRESHOLD = 30;
 
 /**
  * @description: Expandable calendar component
@@ -55,7 +58,11 @@ class ExpandableCalendar extends Component {
     /** whether to have shadow/elevation for the calendar */
     allowShadow: PropTypes.bool,
     /** whether to disable the week scroll in closed position */
-    disableWeekScroll: PropTypes.bool
+    disableWeekScroll: PropTypes.bool,
+    /** a threshold for opening the calendar with the pan gesture */
+    openThreshold: PropTypes.number,
+    /** a threshold for closing the calendar with the pan gesture */
+    closeThreshold: PropTypes.number
   };
 
   static defaultProps = {
@@ -64,7 +71,9 @@ class ExpandableCalendar extends Component {
     firstDay: 0,
     leftArrowImageSource: require('../calendar/img/previous.png'),
     rightArrowImageSource: require('../calendar/img/next.png'),
-    allowShadow: true
+    allowShadow: true,
+    openThreshold: PAN_GESTURE_THRESHOLD,
+    closeThreshold: PAN_GESTURE_THRESHOLD
   };
 
   static positions = POSITIONS;
@@ -82,6 +91,7 @@ class ExpandableCalendar extends Component {
     this._wrapperStyles = {style: {height: startHeight}};
     this._headerStyles = {style: {top: props.initialPosition === POSITIONS.CLOSED ? 0 : -HEADER_HEIGHT}};
     this._weekCalendarStyles = {style: {}};
+
     this.wrapper = undefined;
     this.calendar = undefined;
     this.visibleMonth = this.getMonth(this.props.context.date);
@@ -137,7 +147,7 @@ class ExpandableCalendar extends Component {
   };
 
   updateNativeStyles() {
-    this.wrapper && this.wrapper.setNativeProps(this._wrapperStyles);
+    this.wrapper?.setNativeProps(this._wrapperStyles);
     if (!this.props.horizontal) {
       this.header && this.header.setNativeProps(this._headerStyles);
     } else {
@@ -174,7 +184,7 @@ class ExpandableCalendar extends Component {
         const firstDayOfWeek = (next ? 7 : -7) - dayOfTheWeek + firstDay;
         d.addDays(firstDayOfWeek);
       }
-      _.invoke(this.props.context, 'setDate', this.getDateString(d), UPDATE_SOURCES.PAGE_SCROLL);
+      _.invoke(this.props.context, 'setDate', toMarkingFormat(d), UPDATE_SOURCES.PAGE_SCROLL);
     }
   }
 
@@ -184,10 +194,6 @@ class ExpandableCalendar extends Component {
       return Math.max(commons.screenHeight, commons.screenWidth);
     }
     return CLOSED_HEIGHT + WEEK_HEIGHT * (this.numberOfWeeks - 1) + (this.props.hideKnob ? 12 : KNOB_CONTAINER_HEIGHT);
-  }
-
-  getDateString(date) {
-    return date.toString('yyyy-MM-dd');
   }
 
   getYear(date) {
@@ -204,21 +210,6 @@ class ExpandableCalendar extends Component {
   getNumberOfWeeksInMonth(month) {
     const days = dateutils.page(month, this.props.firstDay);
     return days.length / 7;
-  }
-
-  getMarkedDates() {
-    const {context, markedDates} = this.props;
-
-    if (markedDates) {
-      const marked = _.cloneDeep(markedDates);
-      if (marked[context.date]) {
-        marked[context.date].selected = true;
-      } else {
-        marked[context.date] = {selected: true};
-      }
-      return marked;
-    }
-    return {[context.date]: {selected: true}};
   }
 
   shouldHideArrows() {
@@ -282,8 +273,10 @@ class ExpandableCalendar extends Component {
 
   bounceToPosition(toValue) {
     if (!this.props.disablePan) {
-      const {deltaY} = this.state;
-      const threshold = this.openHeight / 1.75;
+      const {deltaY, position} = this.state;
+      const {openThreshold, closeThreshold} = this.props;
+      const threshold =
+        position === POSITIONS.OPEN ? this.openHeight - closeThreshold : this.closedHeight + openThreshold;
 
       let isOpen = this._height >= threshold;
       const newValue = isOpen ? this.openHeight : this.closedHeight;
@@ -388,20 +381,19 @@ class ExpandableCalendar extends Component {
   };
 
   /** Renders */
+  getWeekDaysStyle = memoize((calendarStyle) => {
+    return [
+      this.style.weekDayNames,
+      {
+        paddingLeft: calendarStyle?.paddingLeft + 6 || DAY_NAMES_PADDING,
+        paddingRight: calendarStyle?.paddingRight + 6 || DAY_NAMES_PADDING
+      }
+    ];
+  });
 
-  renderWeekDaysNames() {
-    const weekDaysNames = dateutils.weekDayNames(this.props.firstDay);
-
+  renderWeekDaysNames = memoize((weekDaysNames, calendarStyle) => {
     return (
-      <View
-        style={[
-          this.style.weekDayNames,
-          {
-            paddingLeft: _.get(this.props, 'calendarStyle.paddingLeft') + 6 || DAY_NAMES_PADDING,
-            paddingRight: _.get(this.props, 'calendarStyle.paddingRight') + 6 || DAY_NAMES_PADDING
-          }
-        ]}
-      >
+      <View style={this.getWeekDaysStyle(calendarStyle)}>
         {weekDaysNames.map((day, index) => (
           <Text allowFontScaling={false} key={day + index} style={this.style.weekday} numberOfLines={1}>
             {day}
@@ -409,10 +401,11 @@ class ExpandableCalendar extends Component {
         ))}
       </View>
     );
-  }
+  })
 
   renderHeader() {
     const monthYear = XDate(this.props.context.date).toString('MMMM yyyy');
+    const weekDaysNames = dateutils.weekDayNames(this.props.firstDay);
 
     return (
       <Animated.View
@@ -423,7 +416,7 @@ class ExpandableCalendar extends Component {
         <Text allowFontScaling={false} style={this.style.headerTitle}>
           {monthYear}
         </Text>
-        {this.renderWeekDaysNames()}
+        {this.renderWeekDaysNames(weekDaysNames, this.props.calendarStyle)}
       </Animated.View>
     );
   }
@@ -436,20 +429,13 @@ class ExpandableCalendar extends Component {
     return (
       <Animated.View
         ref={e => (this.weekCalendar = e)}
-        style={{
-          position: 'absolute',
-          left: 0,
-          right: 0,
-          top: HEADER_HEIGHT + (commons.isAndroid ? 8 : 4), // align row on top of calendar's first row
-          opacity: position === POSITIONS.OPEN ? 0 : 1
-        }}
+        style={[this.style.weekContainer, position === POSITIONS.OPEN ? this.style.hidden : this.style.visible]}
         pointerEvents={position === POSITIONS.CLOSED ? 'auto' : 'none'}
       >
         <WeekComponent
           {...this.props}
           current={this.props.context.date}
           onDayPress={this.onDayPress}
-          markedDates={this.getMarkedDates()} // for Week component
           style={this.props.calendarStyle}
           allowShadow={false}
           hideDayNames={true}
@@ -497,15 +483,12 @@ class ExpandableCalendar extends Component {
             {...others}
             theme={themeObject}
             onDayPress={this.onDayPress}
-            markedDates={this.getMarkedDates()}
             hideExtraDays
             renderArrow={this.renderArrow}
           />
         ) : (
           <Animated.View
-            ref={e => {
-              this.wrapper = e;
-            }}
+            ref={e => this.wrapper = e}
             style={{height: deltaY}}
             {...this.panResponder.panHandlers}
           >
@@ -514,13 +497,12 @@ class ExpandableCalendar extends Component {
               horizontal={horizontal}
               {...others}
               theme={themeObject}
-              ref={r => (this.calendar = r)}
+              ref={r => this.calendar = r}
               current={this.initialDate}
               onDayPress={this.onDayPress}
               onVisibleMonthsChange={this.onVisibleMonthsChange}
               pagingEnabled
               scrollEnabled={isOpen}
-              markedDates={this.getMarkedDates()}
               hideArrows={this.shouldHideArrows()}
               onPressArrowLeft={this.onPressArrowLeft}
               onPressArrowRight={this.onPressArrowRight}