+++ src/expandableCalendar/week.js
@@ -1,9 +1,11 @@
 import PropTypes from 'prop-types';
-import XDate from 'xdate';
-import React, {Component} from 'react';
+
+import React, {PureComponent} from 'react';
 import {View} from 'react-native';
+
 import dateutils from '../dateutils';
-import {parseDate} from '../interface';
+import {parseDate, toMarkingFormat} from '../interface';
+import {getState} from '../day-state-manager';
 import {extractComponentProps} from '../component-updater';
 import styleConstructor from './style';
 import Calendar from '../calendar';
@@ -11,9 +13,7 @@ import Day from '../calendar/day/index';
 // import BasicDay from '../calendar/day/basic';
 
 
-const EmptyArray = [];
-
-class Week extends Component {
+class Week extends PureComponent {
   static displayName = 'IGNORE';
 
   static propTypes = {
@@ -29,78 +29,21 @@ class Week extends Component {
   }
 
   getWeek(date) {
-    if (date) {
-      const current = parseDate(date);
-      const daysArray = [current];
-      let dayOfTheWeek = current.getDay() - this.props.firstDay;
-      if (dayOfTheWeek < 0) { // to handle firstDay > 0
-        dayOfTheWeek = 7 + dayOfTheWeek;
-      }
-
-      let newDate = current;
-      let index = dayOfTheWeek - 1;
-      while (index >= 0) {
-        newDate = parseDate(newDate).addDays(-1);
-        daysArray.unshift(newDate);
-        index -= 1;
-      }
-
-      newDate = current;
-      index = dayOfTheWeek + 1;
-      while (index < 7) {
-        newDate = parseDate(newDate).addDays(1);
-        daysArray.push(newDate);
-        index += 1;
-      }
-
-      return daysArray;
-    }
-  }
-
-  getDateMarking(day) {
-    const {markedDates} = this.props;
-    if (!markedDates) {
-      return false;
-    }
-
-    const dates = markedDates[day.toString('yyyy-MM-dd')] || EmptyArray;
-    if (dates.length || dates) {
-      return dates;
-    } else {
-      return false;
-    }
-  }
-
-  getState(day) {
-    const {current, disabledByDefault} = this.props;
-    const minDate = parseDate(this.props.minDate);
-    const maxDate = parseDate(this.props.maxDate);
-    let state = '';
-
-    if (disabledByDefault) {
-      state = 'disabled';
-    } else if (dateutils.isDateNotInTheRange(minDate, maxDate, day)) {
-      state = 'disabled';
-    } else if (!dateutils.sameMonth(day, parseDate(current))) {
-      state = 'disabled';
-    } else if (dateutils.sameDate(day, XDate())) {
-      state = 'today';
-    }
-    return state;
+    return dateutils.getWeekDates(date, this.props.firstDay);
   }
 
   // renderWeekNumber (weekNumber) {
-  //   return <BasicDay key={`week-${weekNumber}`} theme={this.props.theme} marking={{disableTouchEvent: true}} state='disabled'>{weekNumber}</Day>;
+  //   return <BasicDay key={`week-${weekNumber}`} theme={this.props.theme} marking={{disableTouchEvent: true}} state='disabled'>{weekNumber}</BasicDay>;
   // }
-  
+
   renderDay(day, id) {
-    const {current, hideExtraDays} = this.props;
+    const {current, hideExtraDays, markedDates} = this.props;
     const dayProps = extractComponentProps(Day, this.props);
 
     // hide extra days
     if (current && hideExtraDays) {
       if (!dateutils.sameMonth(day, parseDate(current))) {
-        return (<View key={id} style={this.style.emptyDayContainer}/>);
+        return <View key={id} style={this.style.emptyDayContainer} />;
       }
     }
 
@@ -109,8 +52,8 @@ class Week extends Component {
         <Day
           {...dayProps}
           day={day}
-          state={this.getState(day)}
-          marking={this.getDateMarking(day)}
+          state={getState(day, parseDate(current), this.props)}
+          marking={markedDates?.[toMarkingFormat(day)]}
           onPress={this.props.onDayPress}
           onLongPress={this.props.onDayPress}
         />